mass rename and conversion to typescript
