+++ .circleci/config.yml
@@ -36,7 +36,7 @@ jobs:
             - node_modules
           key: v1-dependencies-{{ checksum "package.json" }}
       - run:
-          name: compile
+          name: compile:all
           command: npm run compile
 
       - run:
