+++ test/persist_banking.js
@@ -6,7 +6,7 @@
 
 var expect = require('chai').expect;
 var ObjectTemplate = require('supertype').default;
-var PersistObjectTemplate = require('../index.js')(ObjectTemplate, null, ObjectTemplate);
+var PersistObjectTemplate = require('../dist/index.js')(ObjectTemplate, null, ObjectTemplate);
 var writing = true;
 /*
 PersistObjectTemplate.debug = function(m, t) {
