+++ test/persist_newapi_tests.js
@@ -46,7 +46,7 @@ describe('persistor transaction checks', function () {
     });
     beforeEach('arrange', function () {
         ObjectTemplate = require('supertype').default;
-        PersistObjectTemplate = require('../index.js')(ObjectTemplate, null, ObjectTemplate);
+        PersistObjectTemplate = require('../dist/index.js')(ObjectTemplate, null, ObjectTemplate);
 
         schema.Employee = {};
         schema.Address = {};
