+++ test/persist_parent_subset.js
@@ -9,7 +9,7 @@ chai.should();
 chai.use(chaiAsPromised);
 
 var ObjectTemplate = require('supertype').default;
-var PersistObjectTemplate = require('../index.js')(ObjectTemplate, null, ObjectTemplate);
+var PersistObjectTemplate = require('../dist/index.js')(ObjectTemplate, null, ObjectTemplate);
 var Promise = require('bluebird');
 var knexInit = require('knex');
 var knex;
