+++ test/persist_schema_indexdefchanges.js
@@ -8,7 +8,7 @@ chai.use(chaiAsPromised);
 
 var Promise = require('bluebird');
 var ObjectTemplate = require('supertype').default;
-var PersistObjectTemplate = require('../index.js')(ObjectTemplate, null, ObjectTemplate);
+var PersistObjectTemplate = require('../dist/index.js')(ObjectTemplate, null, ObjectTemplate);
 
 
 var Employee = PersistObjectTemplate.create('Employee', {
