+++ test/supertype/ExtendedCustomer.ts
@@ -1,4 +1,4 @@
-import {Supertype, supertypeClass, property, Persistable} from '../../index';
+import {Supertype, supertypeClass, property, Persistable} from '../../dist/types';
 import {Customer} from './Customer';
 
 @supertypeClass
