+++ test/supertype/one-to-manychecks.ts
@@ -1,7 +1,7 @@
 
 declare function require(name:string);
 var ObjectTemplate = require('supertype').default;
-var PersistObjectTemplate = require('../../index.js')(ObjectTemplate, null, ObjectTemplate);
+var PersistObjectTemplate = require('../../dist/index.js')(ObjectTemplate, null, ObjectTemplate);
 var logLevel = process.env.logLevel || 'debug';
 
 PersistObjectTemplate.debugInfo = 'api;conflict;write;read;data';//'api;io';
