+++ test/supertype/tsconfig.test.json
@@ -1,9 +1,4 @@
 {
-    "compilerOptions": {
-        "typeRoots": [
-            "../node_modules/@types"
-        ]
-    },
     "extends": "../../tsconfig.base.json",
     "include": [
         "./*.ts"
