+++ tsconfig.json
@@ -1,9 +1,15 @@
 {
     "compilerOptions": {
-        "outDir": "dist",
-        "typeRoots": [
-            "./node_modules/@types"
-        ]
+        "outDir": "./dist",
+        "rootDir": "./"
     },
-    "extends": "./tsconfig.base.json"
+    "extends": "./tsconfig.base.json",
+    "include": [
+        "./**/*.ts"
+    ],
+    "exclude": [
+        "./test/**/*.ts",
+        "./types.d.ts",
+        "./**/persistable.ts"
+    ]
 }
